backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
requests==2.31.0
asyncio==3.4.3

# LangChain and Agent Dependencies
langchain==0.1.0
langchain-community==0.0.12
langgraph==0.0.20

# Ollama integration
ollama==0.1.7

# Redis dependencies
redis==5.0.1
redis[hiredis]==5.0.1

# # Database dependencies (for later)
# sqlalchemy==2.0.23
# alembic==1.13.1
# aiosqlite==0.19.0
# psycopg2-binary==2.9.9
# asyncpg==0.29.0

# For AWS
boto3==1.34.11
aioboto3==12.3.0

# Vector DB and Embeddings
qdrant-client==1.7.3
sentence-transformers
hf_xet

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.1.0